Also persist facts in the hidden lockfile
Otherwise facts are not added to the default lockfile when an entry in the hidden lockfile for a reproducible extension is used.

Also bump the lockfile version out of caution - users may already have hidden lockfiles that lack facts.

Along the way fix the `print` behavior of `Facts` which resulted in `Facts(<unknown object net.starlark.java.eval.Dict>)` since it called the wrong `repr` override.

Closes #27522.

PiperOrigin-RevId: 828559479
Change-Id: I77f002fa0e74c23b9fab5bb05948bcbe76f8f49b