Add executables() query function
This is useful for trying to come up with a list of targets that can be
run with `bazel run`. Transitionally to workaround the lack of this
folks would use `*_binary` to approximate this, but there are many
community rules that don't follow that pattern for things that run
scripts, etc.

RELNOTES[NEW]: bazel query/cquery/aquery now supports an `executables()` function to find only the executable targets in a given expression

Closes #26525.

PiperOrigin-RevId: 828832412
Change-Id: I0bd4a37f788b85d4e02171c11608c8dc99cce305