Support `%bazel_workspace%` interpolation in `--repo_env`
`%bazel_workspace%` is used instead of the more conventional `%workspace%` to avoid changing the behavior of existing `--repo_env` usages in which `%workspace%` may already have a meaning.

RELNOTES: In environment variable values set via `--repo_env`, the substring `%bazel_workspace%` is now replaced with the absolute path of the current Bazel workspace. This can, for example, be used to make tools checked into the repository available on the `PATH` for repository rules.

Closes #25608.

PiperOrigin-RevId: 828862407
Change-Id: I644fbb25d80c5353e522c7184d7c57b2ac8bf04f