[9.0.0] Also persist facts in the hidden lockfile (#27554)
Otherwise facts are not added to the default lockfile when an entry in
the hidden lockfile for a reproducible extension is used.

Also bump the lockfile version out of caution - users may already have
hidden lockfiles that lack facts.

Along the way fix the `print` behavior of `Facts` which resulted in
`Facts(<unknown object net.starlark.java.eval.Dict>)` since it called
the wrong `repr` override.

Closes #27522.

PiperOrigin-RevId: 828559479
Change-Id: I77f002fa0e74c23b9fab5bb05948bcbe76f8f49b

Commit
https://github.com/bazelbuild/bazel/commit/bed7a1f070f27aea2219da7fff5185ecfb71f05e

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>