Only close BES after background tasks have completed
Before this change, the BES would be closed even if remote cache uploads were still happening in the background (due to `--remote_cache_async`, which is enabled by default). This means that the user's Bazel invocation may not have exited while the BES backend is led to assume that it has, resulting in incorrect timing information and potentially also misattribution of cache operations.

Closes #27262.

PiperOrigin-RevId: 828956380
Change-Id: I54808b4f6d67fbf457d1722d6a5a7a7249e977f2