[8.5.0] Add `--remote_max_concurrency_per_connection` (#27564)
The stack traces obtained
https://github.com/bazelbuild/bazel/issues/25232#issuecomment-3462392254
indicate that all gRPC threads are waiting on events when the hang
reported in #25232 occurs, with no other threads being active except for
the virtual threads blocked on upload futures.

This situation is reminiscent of
https://github.com/grpc/grpc-java/issues/8334#issuecomment-892230301 and
further experimentation showed that reducing the maximum number of
concurrent requests per gRPC connection down to 20 (from 100) resolved
the hangs. Reducing the number to 50 made them less likely. Since it is
not clear that there is a single number that avoids hangs for all
backends while not sacrificing performance with some, this change makes
the limit configurable for further experimentation.

RELNOTES: The new `--remote_max_concurrency_per_connection` can be used
to specify the maximum number of concurrent gRPC requests Bazel will
issue on a single connection to the server. The default value of 100
matches the previous behavior.

Work towards #25232

Closes #27466.

PiperOrigin-RevId: 828555281
Change-Id: I901cfb13be7f4f0a4ef1d406845d96e88cecd02f

Commit
https://github.com/bazelbuild/bazel/commit/51a3f0e9bc9d5c2944665ebaf89e0647ac64cd1f

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>