Fix fallout from https://github.com/bazelbuild/bazel/commit/b0f51a9502713e25270a7716bfd4e98a32639c9d.
It added a precondition that asserted that this branch only ever is triggered for LABEL_LIST types, and it just ain't so.

The condition that triggered a crash was a split configuration transition on a label-to-string dictionary attribute.

PiperOrigin-RevId: 829037255
Change-Id: Iddf24302622c33c3ba98e8cb5a7bbbbfe3955fbd
(cherry picked from commit 69b7f883b7e86bb1965faf15c2979caea541f260)