Remove the assertion which disallows stopBranch to be called twice.
Back in the time when this assertion was introduced in 4b1cdfc6df3d662dbc489d21e0ee8d69515e7ee1, it was expected that `stopBranch` should not be called more than once.

However, in the current state, the `stopBranch` function can be called twice when a remote branch wins the race in a dynamic execution, but falls back to a local execution due to issues like downloading error. In that case, the fallback local execution can call `stopBranch` for the second time. We believe this is a valid case and we should remove this assertion to address the issue.

Fixes #27422.

PiperOrigin-RevId: 830370258
Change-Id: Ia8271d8bff66d60596b43d3e908a30ac9e8a5d25