Avoid premature Profiler.instance() calls
...and instead only call it when the profiler is needed.

This is fine when the Profiler is purely static. But with ERW, the instance needs to be set at some point. For production code, we can be reasonably sure that by the time the first Profiler.instance() is called, an instance is already set. Nonetheless, we should instead discourage premature Profiler.instance() calls in class initializations and instead only access it when it's needed.

PiperOrigin-RevId: 830896530
Change-Id: I0d4951a8209c7626957962fde18f46160d66e560