Isolate the Profiler into the SC.
This is a big change, but can be broken down into comprehensible pieces.

**Background**

The trace profiler is used throughout the code base to record the timespan taken by specific parts of the code. It's initialized as a singleton, and is accessed via the static method `Profiler.instance()`. The goal of this CL is to make the profiler available for usage in both the SC and the LC by moving the bulk of its logic to the SC and extracting an interface that the LC code can call.

**Changes**

1/
The existing `Profiler` will be split into the following components and where they eventually go:
- [IJ] `TraceProfilerService`: the public interface of the blaze profiler
- [SC] `TraceProfilerServiceImpl`: the implementation
- [IJ] `Profiler`: the global static accessor of the `TraceProfilerService` instance, which must be set during runtime initialization.

The `CounterSeriesCollector` implementations, which collect peripheral metrics into time series (memory, network, Skyframe nodes) will stay in the LC (except for some usages of JNI, which we'll tackle in a follow-up CL) and is supplied to the Profiler via blaze runtime.

2/
Another consideration is the existing usages of the profiler helper classes outside of the profiler package. An example of this is the `TimeSeries` class. These would also end up in the IJ. Their implementations would be in the SC.

`third_party/bazel/src/main/java/com/google/devtools/build/lib/profiler:{profiler,profiler_impl}` offers a clear view of what goes where.

3/ Non-Bazel code which depends on Bazel code that transitively uses the Profiler without starting it would now receive a `NoOpTraceProfilerService`. There's no behavior changes, as the existing Profiler implementation is also no-op when not started.

PiperOrigin-RevId: 830933793
Change-Id: Ifa35233dce73526d2bce5ca370b938dbec1f3cd7