[8.5.0] Add executables() query function (#27560)
This is useful for trying to come up with a list of targets that can be
run with `bazel run`. Transitionally to workaround the lack of this
folks would use `*_binary` to approximate this, but there are many
community rules that don't follow that pattern for things that run
scripts, etc.

RELNOTES[NEW]: bazel query/cquery/aquery now supports an `executables()`
function to find only the executable targets in a given expression

Closes #26525.

PiperOrigin-RevId: 828832412
Change-Id: I0bd4a37f788b85d4e02171c11608c8dc99cce305

Commit
https://github.com/bazelbuild/bazel/commit/88b5e33fb47f46c3beaccfd9350afe649db4393b

---------

Co-authored-by: Keith Smiley <keithbsmiley@gmail.com>
Co-authored-by: Tobias Werth <twerth@google.com>