Add starlark toml module
This module is analogous to `json` but for TOML. It supports
`toml.decode` and `toml.encode`. This is useful for the increasing
number of places where toml is used over json / yaml, such as Cargo.toml
or uv.lock. Unlike bazel's json support this module depends on a
separate open source project for actually parsing and encoding toml
support. Most of this change is marshaling these objects between
starlark and java representations, which isn't done in the same way
anywhere else I could find.