gen_command-line-reference: store stderr so that it's easier to debug.
This genrule calls the BazelServer binary to perform certain tasks.

One of my recent changes ran into a cryptic failure on BazelCI where this action would fail without anything meaningful to be looked at in the $TMP folder. I had to manually run the command in the action in order to collect the log & debug the issue.

Turns out the error happened because my change caused Bazel to crash due to an NPE (exit code 37) and exit immediately. As a result, nothing was even written in the $TMP dir that could have helped with debugging.

This CL would hopefully save someone the trouble of having to do that in the future, by piping the stderr to a file instead of to /dev/null.

PiperOrigin-RevId: 831823422
Change-Id: Ic74979eaf4607fc1e00a8a0b4dd30731b065fdef