[8.5.0] Support checking out the repo's default branch in `git_repository`/`new_git_repository` rule (#27705)
The `git_repository` and `new_git_repository` rules currently require
not only the remote repository URI to be specified but also either the
branch, tag or revision (commit) to check out. This PR removes the
latter requirement. If no branch, tag or revision are specified, the
repo's default branch will be checked out.

The main reason for introducing this proposed change, is because I
noticed certain Bazel rules already generate `git_repository` targets
from git URIs or other git repo specs that might not have a branch, tag
or revision specified. These rules then typically default to a branch
name of `master`, which is not correct in general. This change will
allow those rules to simply omit the branch, tag and revision from the
generated `git_repository` targets, to check out the repo's default
branch.

An alternative, more explicit, interface for specifying that the default
branch is to be checked out, would have been to add an optional boolean
`default_branch` attribute to the `git_repository` and
`new_git_repository` rules. The presence of that attribute would be
mutually exclusive with the presence of `branch`, `tag` and `commit`.
Let me know if that interface is preferable.

Addresses issue #27610.

Closes #27611.

PiperOrigin-RevId: 833710273
Change-Id: I7bc18efcbcfcaba61bf27052fecceb50336d2eb3

Commit
https://github.com/bazelbuild/bazel/commit/ee3af97622b9d0a0e35be126023835692b629f1c

---------

Co-authored-by: Pieter Agten <pieter.agten@gmail.com>
Co-authored-by: Yun Peng <pcloudy@google.com>