Fix `AbortReason` in `test`/`coverage`/`run` commands.
When BEP completes an event stream it reports an `Aborted` event for all
announced events that have not yet been posted by the build tool. The most
common scenario where this might happen is when the build is interrupted by the
user, or when a build action fails and the tool has been invoked with
`--nokeep_going` (the default).

The `Aborted` events should include an `AbortReason` enum value to provide
context for why the event was not posted. This is particularly important for
consumers to distinguish between crashes, OOMs, user interruptions, and so on.

The `AbortReason` can be determined by the `BuildEventStreamer` through a few
pathways. One important pathway is inferring an `AbortReason` from the
`BuildCompleteEvent` which contains the final status of the build. Critically
this is the only way the `BuildEventStreamer` can determine that the build
failed when the build tool is invoked with `--nokeep_going` (the default).

This handling is more subtle in the case of commands that are conceptually
"build + do something extra": `test`, `coverage` and `run` commands. These
commands produce a `BuildCompleteEvent` as soon as the build portion of the
command is complete, and then produce another `BuildCompletingEvent` when the
"do something extra" portion of the command is complete. Only the build portion can produce incomplete build results.

This change ensures the build tool identifies the `AbortReason` for these
commands when the build portion fails. Previously, for these commands, the initial `BuildCompleteEvent` would be discarded before reaching the `BuildEventStreamer` code that extracts an `AbortReason` from the `BuildResult` in the `BuildCompleteEvent`. Now, the `BuildEventStreamer` will extract the `AbortReason` before discarding the event. (The event is discarded to ensure that it is not handled as a `BuildCompletingEvent` which actually signals that there will be no further events posted.)

PiperOrigin-RevId: 834828863
Change-Id: I0551ef196154e2c9fc18702745b46b9e7cae4987