Add a new optional param to fail() to allow pretty-printing user errors without a stack trace.
This allows a new fail(stack_trace = False, 'Message') which will not include a
stack trace.  This allows an intentional failure that is user facing to be a bit
more user friendly since they're unlikely to want to see the stack trace.  A
flag --force_eval_stack_trace can be used to unconditionally include the stack
trace for debugging.

RELNOTES:
Added an optional parameter to allow eliding the stack trace when calling fail().
PiperOrigin-RevId: 834908566
Change-Id: Ia7fd1673ec90e068b1a23578bd1094c662b94312