Drop discovered inputs immediately after the action in --nokeep_state_after_build mode.
This makes peak heap a little lower and if we don't keep any state after the build, keeping discovered inputs is not very useful.

Sadly, this required quite a bit of BUILD file gymnastics since --keep_state_after_build is in CommonCommandOptions and depending on that in SkyframeExecutor would be a dependency cycle.

RELNOTES: None.
PiperOrigin-RevId: 835127881
Change-Id: I4d1402f662bb9d70f26639f5e782bad1c56d25fb