Skycache: Discard SkyValues after selection
Some SkyValues like PackageValue and ActionLookupValues can be discarded after selection. This removes around 2gb from the heap for a primer build that was ooming.

Regarding the changes to IncrementalInMemoryNodeEntry and the interaction with weakInternPackageTargetsLabels. SkyValues can be discarded but we must keep the nodes with their edges. We set the SkyValue to a new instance of EmptySkyValue. We set them to an EmptySkyValue to distinguish it from null which has its own meaning, a node not yet done. We cannot set it to EmptySkyValue.INSTANCE because that is used during serialization for any empty implementation of SkyValue and therefore has a different meaning too. The instance CLEARED_SKY_VALUE has the meaning that the node was done and it was explicitly cleared.  At the end of the build the InMemoryGraphImpl removes nodes from the graph but before that it adds the package values to a weak interner. By the time we reach the removal step with a cleared node we don't have the package value, therefore we have to add it to the weak interner before clearing.

RELNOTES:none
PiperOrigin-RevId: 835274567
Change-Id: I864bbc6cd62ce71df753e14ab6a80fbc96358436