Remove current_java_runtime from example_test
The fact that the test rule used the `current_java_runtime` toolchain
might account for this CI test error on macOS:

```txt
==================== Test output for //examples/java-native/src/test/java/com/example/myproject:hello:
Error occurred during initialization of VM
java.lang.Error: Could not create SecurityManager
  at java.lang.System.initPhase3(java.base@11.0.15/System.java:2065)
Caused by: java.lang.ClassNotFoundException: allow
  at jdk.internal.loader.BuiltinClassLoader.loadClass(java.base@11.0.15/BuiltinClassLoader.java:581)
  at jdk.internal.loader.ClassLoaders$AppClassLoader.loadClass(java.base@11.0.15/ClassLoaders.java:178)
  at java.lang.ClassLoader.loadClass(java.base@11.0.15/ClassLoader.java:522)
  at java.lang.Class.forName0(java.base@11.0.15/Native Method)
  at java.lang.Class.forName(java.base@11.0.15/Class.java:398)
  at java.lang.System.initPhase3(java.base@11.0.15/System.java:2050)
```

What this is showing is that the `bazel test` command was setting the
correct runtime for the test script, but passing the wrong JVM flags.
The flags come from this expression in `BazelJavaSemantics.getJvmFlags`:

```java
if (JavaRuntimeInfo.from(ruleContext).version() >= 17) {
  jvmFlags.add("-Djava.security.manager=allow");
}
```

`current_java_runtime` returns toolchain info from
`@bazel_tools//tools/jdk:runtime_toolchain_type`. If the selected
toolchain is from the host, that would seem to be the avenue for the
system Java version leaking into this expression.

For the purposes of this test, getting the `java_home` info from the
built bazel should suffice.