ApiExporter: capture missing attributes
There could be special attributes, such as 'package_metadata'
and 'applicable_licenses' defined in lib/packages/RuleClass.java,
that are not part of the exported builtin.pb. This causes language
servers such as starpls to fail diagnosing when encounter these
attributes.

Fix this by exposing ConfiguredRuleClassProvider from SymbolFamilies
and route it down to ApiExporter.collectRuleInfo. There, we include all
the previously unseen attributes into the exported data.

Note that this fix would only add the 'package_metadata' attribute to
the final proto and not 'applicable_licenses' because the latter is
currently an alias of the former.

Part of https://github.com/withered-magic/starpls/issues/410