Automatically retry all blocking system calls on EINTR.
This is required for robustness against signals, but we didn't consistently do it.

It's not always clear which system calls can return EINTR (it varies across OSes and filesystems) so to be safe I've wrapped all of them, which should harmless to the ones that don't need it.

An important exception is close(2) (and, indirectly, closedir(3)), for which retrying on EINTR is dangerous as it might cause a different file to be closed. (POSIX says that the file descriptor is left in an unspecified state, but all modern Unix systems close it before returning EINTR.)

In addition, converting EINTR to InterruptedIOException seems inappropriate: the latter specifically means that the calling thread was interrupted (in the Java sense), which has nothing to do with signals.

PiperOrigin-RevId: 842259373
Change-Id: I81401db0c7e939248a589221403a7d15a3fac639