[9.0.0] Add support for multiple module interfaces per `cc_library` (#27927)
When multiple `module_interfaces` are specified on a single
`cc_library`, the individual compilation actions form a DAG based on
`import`s between these modules. Consider the following situation:

* `a.cppm` imports `b.cppm`, both of which are in the
`module_interfaces` of a single `cc_library`.
* Building the target populates the action cache with an entry for
`a.pcm` that stores `b.pcm` as a discovered input.
* Now edit `a.cppm` and `b.cppm` so that `b.cppm` imports `a.cppm` and
`a.cppm` no longer imports `b.cppm`.
* Build again (optionally after a shutdown).

Before this commit, this resulted in an action cycle since during action
cache checking, Bazel would reuse or look up the inputs discovered in
the previous build, thus introducing an edge from `a.pcm` to `b.pcm`.
Together with the newly discovered edge from `b.pcm` to `a.pcm`, this
resulted in a cycle.

This is fixed by not requesting the previously discovered inputs (either
retained in memory or in the action cache) if the mandatory inputs
changed. In the case of C++20 modules, this is sufficient since the
modmap file, which lists all transitive `.pcm` files required for
compilation, is a mandatory input.

As part of this change, `MetadataDigestUtils.fromMetadata` had to be
modified to always return a byte array of proper digest length, even if
called with an empty map, to match the assumptions of the action cache.

This change is pretty much Fabian's PR
https://github.com/bazelbuild/bazel/pull/27492 with a tiny fix added on
top (not returning from computeMandatoryInputsDigest() early on
valuesMissing() if inErrorBubbling() is true)

Closes #27492.

PiperOrigin-RevId: 842733471
Change-Id: I48fa2c0bceb888dcb58db29d50c30719b2122c5d 
(cherry picked from commit cb9bd8615210dda2104f79d281938e47187dc2de)

Closes #27544