Fix type of resolved function, add tests
This adds tests to `TypeResolverTest` for a resolved function's type, and fixes a couple bugs:

- `**kwargs` should not be included in `getParameterNames()` / `getParameterTypes()`
- `getVarargsType()` / `getKwargsType()` should be `null` rather than `Types.NONE` when the `*args` / `**kwargs` parameter is absent.
- Make sure to recurse into the bodies of `def` and `lambda` when assigning types. (This is probably a regression from when we factored out typing code from `Resolver` to `TypeResolver`.)

`TypeResolver#resolveFile()` is renamed to `annotateFile()` to help better distinguish it from symbol resolution. Note that its job is to evaluate type annotations and apply the resulting `StarlarkType`s to the AST, so it is "annotating" the AST. Maybe there's another name that's clearer still...

Work toward #27728.

PiperOrigin-RevId: 843205294
Change-Id: I46cfa10b78d110c4fb127e61ee2c13939fc43f8f