list_source_repository: fix source not updating after deletion
Tell the repo rule to watch the workspace and rerun would be the best
way for us to get accurate data.

The repo rule is only used for //src/test/shell/bazel:srcs_test so users
who don't run this test would not have to re-run the repo rule for each
change.

Closes #27938.

PiperOrigin-RevId: 843579713
Change-Id: I35446c9ceabc132b9fa314c73522c4ebb08aa51c