`TypeResolver` stores type info in `Binding`
Resolver bindings gain a new field to hold type information, initially set to null but then populated by the `TypeResolver`. This mirrors how resolved function objects work.

A null type is treated as `Any`. We are inconsistent about which one is used for an unannotated symbol -- `Any` is used in function parameters to match the behavior for resolved function types, and `null` is used elsewhere. Supporting null in this field is useful for treating untyped code as `Any`.

In principle an AST should be resolved only once, but possibly put through the type resolver any number of times (to support redetermining type information with modified type definitions, when user-defined types are present). We require the AST to have null types for all its bindings and resolved functions before running the type resolver. This precondition ensures we aren't inadvertently overwriting type information. An extra pass that nulls out types should be done between multiple type resolver runs, both to satisfy this precondition and to ensure any information from the prior run isn't accidentally carried forward to the new run.

A side effect is that the type resolver now rejects programs that attempt to redefine the same function (e.g. when reassignment of top-level names is allowed, or when multiple defs are given for the same inner function). We have to tolerate such programs in untyped code for backwards compatibility. In typed code, we previously aimed to allow this in some narrow cases for symmetry with how other identifier bindings are handled, but it seems simpler and more sensible to just disallow it outright. A follow-up CL will move enforcement of this from `Resolver` to `TypeResolver`.

Fixed `Program#compileFile()` to run the `TypeResolver` iff type checking was requested. This means threading `resolveTypeSyntax()` through `BzlCompileFunction`. Updated `StarlarkTypesTest` to ensure we're getting the right resolver / dynamic type checking behavior for the right Bazel flags. Updated `DynamicTypeCheckingTest` to not use unknown symbols in type annotations.

Add helper functions to `TypeResolverTest` to improve readability.

Work toward #27728.

PiperOrigin-RevId: 843705833
Change-Id: If773dd1724c093f992743b0dab271b6692953bf0