[9.0.0] Add type param to repository_ctx.extract() (#27923)
Before, you had to rename or symlink a file to the correct
compressed/archived file extension in order for it to be decompressed
properly. Now you can specify rhe type of the file explicitly, without
renaming/symlinking.

In other words, before, this wouldn't work, because the file type could
not be determined:

> repository_ctx.extract("my_archive")

Now you can set a type (the archive is zipped):

> repository_ctx.extract("my_archive", type="zip")

Fixes: #26254

---

The first commit does some refactoring before adding the feature. This
change will merge conflict with #27413 - I'll update once that is
merged.

Closes #27423.

PiperOrigin-RevId: 842655968
Change-Id: Iea9e3fd458229a7cdd8f6a076eb594d82d61b5b0

Co-authored-by: Will Stranton <2659963+willstranton@users.noreply.github.com>