[9.0.0] Fix --starlark_cpu_profile time misattribution.
Previously, we flushed the per-thread tick counter when exiting a stack frame. This resulted in ticks already accrued by frame A before it calls frame B to be misattributed to B.

To fix it, also flush the counter when entering a stack frame.

The issue may be observed by profiling this rule:

```
def inner():
  x = 0
  for i in range(100000000):
    x += 1

def outer(ctx):
  x = 0
  for i in range(100000000):
    x += 1

  inner()

  x = 0
  for i in range(100000000):
    x += 1

  return DefaultInfo()

burn = rule(outer)
```

You'd expect one third of the time to be spent in `inner` and the remaining two thirds in `outer`. However, we currently report one third spent in `range`, one third in `inner`, and the final third in `DefaultInfo`.

PiperOrigin-RevId: 845199672
Change-Id: I5d9ffc2e5a779c96ce64c1bf77e8b84429c454e3