Limit the number of in-flight SkyValues during serialization.
The limit is currently a constant 10K. I wanted to make it configurable by a command line flag but it turned out that my initial guess worked pretty darn well and plumbing command line options to this place is not trivial, so I let it be for the time being.

RELNOTES: None.
PiperOrigin-RevId: 845308180
Change-Id: I98c524615df3f2c12686fbe05b4331c8c497242a