[9.0.0] Support `{name}` and `{version}` placeholders in `use_repo` (#28016)
This allows module extensions to reliably namespace repos created by
individual modules, even in the presence of `multiple_version_override`,
without causing additional churn for users during version bumps. In
particular, projects with a publishing workflow that patches in the
`version` attribute of the `module` function don't need to modify that
workflow to e.g. patch a global constant or also search for the old
version in `use_repo` call arguments.

Context:
https://bazelbuild.slack.com/archives/C09E58X3AQ7/p1764934858912079

RELNOTES: The values of keyword arguments passed to `use_repo` can now
contain the special substrings `{name}` and `{version}`, which are
treated as equivalent to the corresponding attributes of the current
module.

Closes #27890.

PiperOrigin-RevId: 845156030
Change-Id: Ic5d777645bf6a5ca67fa6cb577f426c96beda894

Commit
https://github.com/bazelbuild/bazel/commit/7ca1159de8643ddffc781f00780f61f8d3e9c502

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>