[9.0.0] Preserve order of recorded inputs (#27985)
Recorded inputs for repo rules and module extensions are no longer
separated by type, but recorded in the order they are encountered. This
is necessary preparatory work to avoid Skyframe cycles that could
otherwise occur while checking cache, lockfile or marker file entries
for up-to-dateness.

This requires a breaking change to the lockfile structure, but has the
positive side effect that each recorded input is now represented as a
single line in the lockfile and thus more compact.

Along the way, the error message for an out-of-date lockfile entry with
`--lockfile_mode=errors` are improved and now include the precise reason
for the invalidation.

For repository rules, the repo mapping entries recorded while loading
the repo rule definition are now included in the predeclared inputs
hash, which makes cache entries more specific and avoids unnecessary
Skyframe lookups.

Work towards #27517

Closes #27603.

PiperOrigin-RevId: 844721202
Change-Id: Id063a88508591086b07bd89a822cf22b2d90610d

Commit
https://github.com/bazelbuild/bazel/commit/41ccfefb88e9d816453eb42e549895b179c64061

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>