Serialize SkyKeys for ACS lookup asynchronously.
Otherwise, a deadlock is possible: serialization can submit tasks to ForkJoinPool.commonPool() to do compression (see UploadOnceFuturePutsResolve), but we call submitInvalidationLookup in the same pool. These would then wait for UploadOnceFuturePutsResolve to complete the serialization, but the latter cannot make progress since there are no threads in the pool to handle the aforementioned tasks they submit because all of them are blocked in submitInvalidationLookup().

The fix is to not block on serialization in submitInvalidationLookup().

RELNOTES: None.
PiperOrigin-RevId: 845771785
Change-Id: I649a28d9193e86a681dccca0adae22b30b94988e