Refactor baseline CL change tracking to occur earlier.
Notably, we now know the baseline diff early enough to still reset the evaluator. This will be used to implement a feature to start with a fresh graph if the diff exceeds a certain threshold.

`DiffAwarenessManager` is modified to cache the current `View` between calls to the new method `getEvaluatingVersionDiff` and `getDiff`.

PiperOrigin-RevId: 846733217
Change-Id: Ib65b9c86c8c9e7e302acd9b3ea2a5f91ebbde2f2