Enable `--experimental_retain_test_configuration_across_testonly`
This avoids action conflicts when `testonly` non-test targets depend on test targets.

Until `--incompatible_check_testonly_for_output_files` is flipped, it is still possible to get action conflicts when a non-`testonly` target depends on an output file of a test. Since this can be fixed by adding the missing `testonly` annotation, this is only a minor concern.

Fixes #28056

Closes #28073.

PiperOrigin-RevId: 847892897
Change-Id: I60e908a22b51c307306760aa054b3968cfbebac2