Optimize builtins restriction check
`BuiltinRestriction` currently accounts for 9% of the total CPU time for `bazel build --nobuild //src:bazel-dev`. This is avoided by replacing the linear search with a sorted map lookup and avoiding the repeated creation of "non-visible" `RepositoryName` objects, which trigger a `SpellChecker` computation.

The map lookup is made possible by the transition to Bzlmod, which makes it possible to drop legacy repo names and treat the remaining apparent repo names as module names, for which it is easy to derive a prefix of the canonical repo name.