Use feature instead of cpp fragment for dSYM generation
This was discussed in https://github.com/bazelbuild/bazel/pull/25881
where we believed it was equivalent. This isn't equivalent if the
toolchain does not support dSYM generations but --apple_generate_dsym is
still passed (as is possible on Linux). In this case we let feature
configuration handle this instead, as this feature is requested when
--apple_generate_dsym is passed. In that case because of this mismatch
this directory will be created "successfully" but it will always be
empty (this would be a failure if it wasn't a tree artifact)

Closes #26925.

PiperOrigin-RevId: 848240203
Change-Id: I604bc97e44c195142b12222b73d7442a1115ccd1